Fixed issues and added pause for bones
Pausing with spacebar now pauses the bones as well.
Tidied up script and removed bone void.
Fixed issue where line was not displaying. This was because of noStroke(). Instead the stroke is set to match the fill of the bone and then reset at the end of the display function.